kafka/tx: support invalid_txn_state error code
Redpanda protocol depends either on calling abortTransactions on a failure
during adding a partition to a transaction

    try {
        producer.beginTransaction();
        producer.send(...);
        producer.send(...);
        producer.commitTransaction();
    } catch (KafkaException e) {
        producer.abortTransaction()
    }

or closing and recreating a producer.

It happen that Kafka client doesn't always send end_txn RPC to a server
when .abortTransaction() is invoked.

See https://issues.apache.org/jira/browse/KAFKA-5260

By returning INVALID_TXN_STATE we'll force a user to close and recreate
a producer.