admin brokers command: remove describe, remove stray fmt.Println
The "describe" command for brokers only existed because at the time the
list API did not return the membership status.

Now that list returns membership status (#1933), describe is useless: it
returns the same information as list, but only for one broker.

This also removes a stray println that accidentally made it into tip,
as well as a stray `a.Brokers(0)` that was used at one point for
testing.