kafka/tx: make end_txn flow resistant to exceptions
We use early return by passing a promise (an outcome) to a function when we
process the end_txn requests. Since we process the "true" function result after
processing the outcome it may lead to the stalls when the function fails with
an exception before setting the outcome.

We don't use exceptions in the consensus layer so it should not happen but it's
better to be sure. So adding try/catch error handling to prevent this case.