c/members_manager: apply raft0 configuration from stm
In members_manager we process configuration updates from raft0 as soon
as they are appended to raft0 log. This processing is made in background
and is not synchronized with other controller STM state updates. It may
happen that topic updates being older than configuration update will be
applied earlier. This may lead to an error in which the state of
partition allocator will not be correctly updated. Fixed issue by
applying raft0 configurations to members_manager state. This way we
guarantee that configuration updates will execute in order with regards
to topic state updates.

Signed-off-by: Michal Maslanka <michal@vectorized.io>