c/topics_table: validate if finish command matches update command
Added validation checking if received `finish_update` command matches
previously received `update` command. If this is not the case do not
generate delta and return an error. Previously `finish_update` command
related with different update was finishing different update which was
causing errors in controller backend and may lead to situation in which
partition would be removed before it is fully replicated to target
nodes.

Signed-off-by: Michal Maslanka <michal@vectorized.io>