raft: reset RPC backoffs on prevote from unresponsive follower
This is an optimization, to encourage leaders to send
heartbeats more promptly after follower restart.  This case
triggers when a restarted node hasn't received heartbeats
promptly, and sends prevotes - the prevote induces the leader
to reset their backoff.

This could be improved further by having nodes emit
a 'hello' message as soon as they start, rather than
waiting for the restarted node to give up and start a vote.

Signed-off-by: John Spray <jcs@vectorized.io>