raft: don't approve prevote requests when leader heartbeats are fresh
This change is intended to improve leadership stability.

Previously, if a restarted node didn't receive an append_entries_request
almost immediately on startup, and the partition was idle, they would enter an election
pass their prevote check because their log is up to date, and in
the process cause the existing leader to step down.

Now, if a majority of nodes have had recent heartbeats from the leader,
they will refuse prevotes and thereby prevent other individual nodes
from disrupting the cluster.

Ideally, the prevote success condition should be the
same as the real vote (i.e. check `granted`).  However,
because we emit prevotes with the old term, they will
always have granted=false, and if we sent prevotes with
an incremented term, they might cause leaders to step down
too early.  A more thorough rework of prevote is a job for
another day.

Fixes: https://github.com/vectorizedio/redpanda/issues/2048

Signed-off-by: John Spray <jcs@vectorized.io>