r/consensus: fixed suppressing follower heartbeats
In redpanda raft implementation we allow to have multiple append entries
request pending in flight to each follower. When append entries request
is send to the follower we suppress sending heartbeat requests. When
multiple requests are pending to be replied by the follower we have to
guarantee that heartbeats will be suppressed until the last request was
replied.
Fixed updating heartbeat suppression state by correctly using MVCC in
follower stats. Now we only allow to update heartbeat suppression state
with response that is later than last update. This fixes not necessary
log truncation that was occurring when testing redpanda with heavy load.

Signed-off-by: Michal Maslanka <michal@vectorized.io>