cluster offsets: switch to franz-go
This builds on work introduced in previous commits.

Unfortunately, there is a lot more rote work to do for admin commands
when using franz-go. The end result is we are more resilient to broker
failures.. The end result is we are more resilient to broker failures..
The end result is we are more resilient to broker failures. The end
result is we are more resilient to broker failures, have better control
over when to bail, etc.

This changes the output slightly in three ways:
- we are now more in line with kafka's own kafka-consumer-groups.sh for
the naming of headers
- we separate groups with a leading header block that includes
information that was previously never printed (this also allows us to
print empty groups)
- the LAG % column is dropped: the percentage becomes increasingly
meaningless as the number of messages written grows.

The test is dropped as well. franz-go does not have the concept of
mocking yet, nor does it provide an easily faked kafka for testing
against. The existing test simply tested some golden output; I think
golden output tests can be useful in some scenarios, but is less useful
in CLIs where we want to constantly improve the output.

I've ran this command against a local cluster with
- no groups
- two groups, both active
- three groups, one dead