r/vote_stm: using timeout when replicating configuration after becoming a leader
It is possible in some situations that node will have to wait
indefinitely for configuration to be replicated after it became a
leader. (consider situation in which there is only one voter in raft
configuration). If this situation happen vote timeout will never be
rearmed and no new leader will be elected in the cluster.

Example that may lead to the error:

Raft group configuration (voters: {1}, learner: {2})

1) node 1 became a leader
2) node 1 request replication (request to {2} was skipped since it is
   recovering)
3) node 1 receives append_entries_request from a node that is replying
   controller history (old revision of current raft group)
4) node 1 lose leadership
5) node 1 waits for replication to finish - vote timer is not rearmed

In this scenario node 2 can not be elected as a leader since it is a
learner. There is no leader in the raft group

Signed-off-by: Michal Maslanka <michal@vectorized.io>
(cherry picked from commit b9cc0307358479f510c367a6d690849b7fe00661)