cluster: calculate and use muted nodes in leader balancer
To determine if a node should be muted we use a simple heuristic of
expiring heartbeat on raft-0 group. A node will be considered muted if
raft0 has not received a heartbeat message within some time interval
(e.g. 20 seconds). Once heartbeats resume we reconsider the node as
available capacity for accepting leadership transfers.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>