cluster: add muted node concept to greedy balancer
Muted nodes are not considered for available capacity when balancing.
This is useful when we want to consider a node to be offline, and
balance among the other nodes. The error calculation does not consider
muted nodes so that error reporting represents offline cores, which can
be useful when we see that there are no improvements but a large error
value.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>
(cherry picked from commit d7bdb045003cd71dd0b816dda3c21ec7ac20488b)