ts-generator: fix missing parameter types
In TypeScript, the parameters of a function signature must
have a name and optionally a type. Missing types on parameters
will default to any since no type was given.
Functions using that signature will not emit an error on compile time
only on run time. The `FromBytes<T>` function signature is
an example of this issue.

This patch increases the type-safety by adding the correct types
for the `FromBytes<T>` function signature.

Signed-off-by: Dugagjin Lashi <lashi.dugagjin@gmail.com>