r/consensus: protect start with gate
When consensus is being started it access an underlying `storage::log`.
We have to start consensus under the gate since `start()` may race with
`stop()`. In this case we have to wait for start to finish and then
finish cleaning internal state. The race between starting and stopping
raft is caused because creating raft groups and shutting down redpanda
is not synchronized. In future we have to refactor startup & shutdown
sequence to protect such kind of races.

Signed-off-by: Michal Maslanka <michal@vectorized.io>
(cherry picked from commit 9f13de8e8aae1725750c68a2de4363fc27f37571)