cloud_storage: Add partition_recovery_manager
The component will be used by the partition_manager to recover
partitions that have 'x-redpanda-recovery' override set. The component
can be used to download log segments for the particular ntp. The
download process takes into account the retention policy of the
topic:
- If 'retention.bytes' is used it will download last 'retention.bytes'
  of the log (aligned to segment boundary). It cant download more data
  than the policy allows.
- If 'retention.time' is used it will download latest log segments.

The recovery process is transactional. If the node is failing during
recovery the recovery process will continue from the last fully downloaded
segment. If the bucket content is corrupted (e.g. some manifests or
segments are missing) the recovery process will still succed. It will
attempt to fetch what's possible. The user might delete the topic and
re-run the recovery process if needed. Note that from the recovery
perspective it's not possible to say if the partition data was delted
from S3 or it was never uploaded.

If the cloud storage is not available (downloads are timed out) the
recovery will fail and will be retried later. The partition is not
available for read/write during recovery. User will never see the
partially recovered data.

This commit also implements segment filtering using offset_translator.
Every log segment is parsed and raft_configuration batches are removed.
Also, the file name of every segment and offsets of some batches are
also changed accordingly.

Also, this commit adds extra default override to ntp_config that's used
by the partition_recovery_manager.