persisted_stm: make sync don't affect offsets
sync used to write a checkpoint to discuver the latest
replicated offset and then waited until the corresponding
event is applied.

From the client's perspective the checkpoint event looked
like a gap so the offset were shifted. Making sync not
to write to disk to preserve the default sequence of the
offsets.

fixes https://github.com/vectorizedio/redpanda/issues/2111