schema_registry/seq_writer: Don't retry non-retryable errors
sequenced_write should only retry on a collision.
sharded_store returns exceptions for errors that are not retryable.
retry_with_backoff retries if there was an exception.

Wrap the result of f passed to sequenced_write_inner in an outcome, so
that errors reported by the store are not retried, unwrap the outcome
in sequenced_write.

Signed-off-by: Ben Pope <ben@vectorized.io>
(cherry picked from commit 0cdebaabff3fb81db742331407cd9bcfd0c4592e)