utils: add QuDiTracker
This wraps the common pattern of waiting for some
resources and then dispatching an inner operation: we
track the queuing time/depth and the dispatch time/depth
independently, so that an operator can understand better
the source of their latency.

Because queue depths can vary quickly, that is also tracked
in a histogram to make the distribution of depths apparent.

Signed-off-by: John Spray <jcs@vectorized.io>