kafka: add EachShard for franz-go shard output
The pattern introduced by the conversion to franz-go in `rpk cluster
offsets` is to print messages when individual shards fail, and to fail
when all shards fail.

This abstracts the individual shard failure logging, and returns whether
all fails. The rpk cluster offsets command will be converted to use
this, and then this will be used in switching up rpk topic describe.