rpk topic: convert describe to franz-go, remove & alias info
The new version of the describe command is more resilient to individual
broker errors, and includes more information.

The info command has been removed and made an alias for describe, all
previous flags still exist but are marked deprecated; this allows us to
avoid hard failing on existing users. We can remove these flags when
removing previously deprecated flags in the future (i.e. in a few
months).

The previous describe command had the concept of printing sections of
partitions at a time, allowing a person to print 1 thru 20 of 100
partitions. The default was to print everything at once. As well, the
default was to include all watermarks.

After pinging Roko, the page aspect may not be used so much, and
following the principle introduced in the rpk cluster metadata command,
we can leave it to users to pipe output to a file / to less so that they
can filter as they need. It's quicker and easier.

However, one change with this command is that we no longer print
partitions by default. We now only print the summary section and the
configs section. Detailed partition information can be opted into with
-a or -p. We now also print the log start offset & high watermarks
(previously we only printed high watermarks). If the last stable offset
is different from the high watermark, we print that too (this is
relevant for transactions).

As well, in the config section, we now always print all configuration.
There aren't actually that many configuration options for a topic, and
it can be quite relevant to look at all configs, even default configs,
when we describe a topic.

Finally, the info command is removed because was effectively describe
with just the summary section. The under replicated portion was not
relevant right now, because redpanda currently always returns all
replicas as in-sync all of the time.

This replaces the test for the describe command with a test ensuring the
headers / rows of partitions have the appropriate arguments that will be
printed. This is more resilient to changing individual bits of text if
we want to when we format the output, and will be easier to fix going
forward. The other aspects of printing are very small, tests would not
be beneficial here.