resources/smp: changed maximum number of concurrent x-core request for raft
Increased the maximum number of concurrent x-core dispatches in raft
group. Previous value of 5000 imposed an artificial limit on number of
requests that redpanda could handle. Requests queued in `raft::service`
were waiting on available slots in `raft` smp service group which is a
waste of time since they can be dispatched to the group and handled in
batch.

Signed-off-by: Michal Maslanka <michal@vectorized.io>