cluster: Deletion of materialized & source topics
- A source topic to a materialized topic cannot be deleted if there are
one or more materialized topics that are currently based off of it.

- A materialized topic may be deleted at will, no lookups are performed
within v/coproc to ensure there aren't coprocessors that are writing to
this topic, as materialized topics are 'auto-created' and can be created
on demand.

- Upon processing of the delete command the relevent data structures
must be updated, i.e. shard_table, partition_leaders_table must have any
in memory references to the materialiezed topic/ntp removed and
storage must removes the on disk log segments themselves.