rpk topic list: convert to franz-go
This conversion actually relies on something that @SteadBytes added in
kcl: we can just call the metadata command with the -t flag.

In rpk we also add the -i flag.

The main difference in output is that we by default do not print
internal topics now, and the headers are uppercased. No other
difference. The list command now supports -d to print per partition
information.