rpk topic create: convert to franz-go
In this conversion, we also add the ability to create multiple topics at
once.

The key/value parsing has been changed to also accept `=` to separate
the key and value, and all help text uses '='. This is more in line with
kafka-configs.sh, and more in line with how we use key=value for
everything else (env vars, flag key/value separators).

This deprecates the --compact flag but still supports it, printing that
the user should use -c cleanup.policy=compact. This avoids compact being
special. This also avoids --compact=false (i.e., the default) explicitly
setting compact.policy=delete, when instead we want to rely on the
broker default.

We do now provide setting compact as an example in the help text.

As mentioned, we avoid sending the compact policy if it is not
specified, because brokers themselves have a default and we shouldn't
override that.

There's very little difference in the output text; wheras previously we
just said "Created topic foo", we now use column output, which is better
for the new support of creating multiple topics at once.

Lastly, this also adds the --dry (-d) flag, which may be useful
sometimes for some people in some scenario someday.