topic test: remove
TestTopicCmd previously tested list, create, delete, and set-config.
List is now provided directly by metadata, and create & delete are
extremely simple requests that do not really have much to test (mostly,
these were tests that the mocks returned things).

set-config has useful tests, but I'm currently in the process of
changing set-config a little bit to be incremental. I'll explicitly test
that a request is created as expected for set-config specifically in the
next PR.