storage: Add transfer_stream function
The function can be used to pipe an input_stream to the output_stream.
It's similar to seastar::copy in this regard. But this fucntion can
also transform the stream using the user supplied lambda function.
The function is invoked per record batch. It receives a mutable
reference to header that was read from the input stream. The function
can modify the header. The moidfied header will be written to the
output_stream instaed of the original one. Also, the function can
stop transfer process or skip the record batch.