archival: Switch from high watermark to LSO
Make archival subsystem look at last stable offset instead of high
watermark. Because of transactions high watermark can point to the place
in log that might contain data that can be roled back eventually.

Also, the index might return the offset which isn't flushed yet. This
commit implements the workaround for this situation.