s/log_reader: made hole checking logic independent from cleanup policy
Updated logic checking if holes were encountered in a log to be
independent from cleanup policy. When cleanup policy changes from
`compaction` to `deletion` holes may exists in a topic that is not
longer compacted. Relaxed hole checking logic to still validate if
offsets are monotonic but made it independent from cleanup policy.

Fixes: 2214

Signed-off-by: Michal Maslanka <michal@vectorized.io>