rpk topic: set-config redux; rename to alter-config
This is the first true breaking change in the series of commits
converting rpk to using franz-go, but it's quite worth it.

Our previous set-config implementation used the original AlterConfigs
API, which loses any dynamic configuration any time an alter happens if
the key is not set.

The new implementation uses incremental altering. The syntax is
fundamentally different, so now we require --set, --delete, --add, and
--remove flags. We can also set configuration for multiple topics at
once.

All heavy lifting is done by parseKVs and the kgo.Client. The logic of
incremental altering is only building a request and issuing it.

The command itself has been renamed to alter because we now (a) support
more than just setting configuration (we now can delete), and (b) this
more accurately mirrors the request / any kafka documentation.