archival: add cloud_storage cache service
Shadow indexing is a project to retrieve data from S3 storage in a
transparent way for user. Fetching data from S3 will create two data
streams, one of which will be passed to the client, and another one will
be passed to the cache system. When retrieving data, first cache will be
checked, and if data is not present in the cache, then it will be
retrieved from S3. Both segments and manifests will be cached.