utils: Add stream utils
Add the 'input_stream_fanout' template that takes a single seastar
input_stream instance and returns multiple.

Number of returned input_stream instances depend on template parameter:

  auto [s1, s2, s3] = inpust_stream_fanout<3>(std::move(s, 4);

The function also accepts 'readahead' and 'memory_limit' parameters.
The 'readahead' limits number of buffers that can be in-flight at the
same time. The 'memory_limit' limits total amount of memory used for
buffers which are in-flight at the same time.

It is possible to crate from 2 to 10 streams. The high end limit of
10 is not technically a limit. It is possible to create more. We just
don't need that.

Internally, the fanout input stream has a collection of in-flight
buffers. The background fiber constantly reads new buffers from the
supplied input stream and adds these buffers to the collection. Every
buffer has associated mask and semaphore unit. The semaphore units are
acquired before reading from the stream. The max-count property of the
sempahore is equal to 'readahead' value.

The cosumers are scanning the collection from the begining. Every
consumer has a unique id and associated bit in the mask. If this bit
is not set the consumer will consume the buffer (by sharing it) and
set the bit. If the mask after that is full (all bits are set which
means that the buffer is consumed by all consumers) the consumer will
remove the buffer from the collection.

If the collection is full the producer waits until the consumers will
free up some space. If the consumer can't find an element to consume
because either the collection is empty or it's fully consumed by that
particular consumer, it will wait on the condition variable. The
condition variable is trigered on produce.