admin: fix GET on partition details for controller topic
Previously we included the controller topic in the list
of topics, but a GET on the /partitions/redpanda/controller/0
would return 404.  This behaviour was inconsistent, and also
makes it awkward for any automation/testing that wants a single
API endpoint to use to find topic location/leaders.

Today, then controller topic is replicated on all nodes, but
in future we expect to change to keeping it on a subset of nodes
in order to better support larger system sizes -- at that time
the controller topic would be managed more like a normal topic
and the special casing here would go away.

Fixes: https://github.com/vectorizedio/redpanda/issues/2294

Signed-off-by: John Spray <jcs@vectorized.io>