r/utils: fixed filling gaps with size larger than max batch size
Raft protocol doesn't tolerate gaps in log offset. In order to make
recovery of compacted topics possible we use ghost batches (virtual
batches that are not persisted to disk). Previously we always used
single ghost batch per gap. If the gap was large enough it may lead to
situation in which `batch_header::offset_delta` field was overflown.
This may lead to crash on the node receiving append_entries requests
since offsets are not monotonic, and assertion is triggered.

Fixes: #2310

Signed-off-by: Michal Maslanka <michal@vectorized.io>