r/consensus: add all previous configurations to manager when starting
Since we are using `raft::configuration_manager` as an index of redpanda
control batches involved in offsets translation we need to make sure
that log content matches the manager content. It is not enough to add
only the last seen configuration into the manager since it would corrupt
offset translation.

Signed-off-by: Michal Maslanka <michal@vectorized.io>