r/config_manager: make prefix truncate to preserve last configuration
When prefix truncating configuration manager it is possible that the
truncation point will have an offset greater than the last stored
configuration. In this situation simply removing configurations from a
map stored in a manager is not an option since it would lead to
information loss. Fixed prefix truncate operation not to require
configuration to be present with offset larger than truncation point.
Instead now the configuration manager moves last known configuration to
the truncation point. This operation preservers configuration and
indexing information used during offset translation.

```
                          200
                           │
                           │
                           │ prefix truncate
                           │
 ┌───────┬───────┐         ▼
 │   0   │  100  │...........................
 └───────┴───┬───┘
             │              ┌───────┐
             └─────────────►│  200  │...............
                            └───────┘
                    move last know configuration to truncate offset

NOTE: box with number represent an entry in configuration manager

```

Signed-off-by: Michal Maslanka <michal@vectorized.io>