r/consensus: do not add configuration to manager when storing snapshot
Configuration manager serves as an index of metadata batches for Kafka
offset translation. Inserting configuration with offset smaller than the
last one into the `raft::configuration_manager` will lead to the situation
in which all the offset translation deltas for kafka offsets following
the offset of inserted configuration will change (there will be more
control batches before those offset since we inserted one)

This change in offset translation breaks the WAL immutability property
and may lead to the situation in which some of the batches are
lost by consumers. Batches are in the logs but since offsets were moved
consumer just skipped it over.

Signed-off-by: Michal Maslanka <michal@vectorized.io>