raft: make stable leadership behaviour stronger
When a restarted node is not yet receiving append_entries
messages, it starts a prevote.  Previously this prevote would
tend to succeed, if not on the leader then at least on a
follower, encouraging the restarted node to increment
its term and displace the leader.

The stable leadership optimization makes followers
refuse votes if they have heard from their leader
recently (and makes the leader refuse votes generally).  By
moving it earlier in the do_append_entries function, we
avoid not just voting for peers, but passing their
prevotes as well.

Signed-off-by: John Spray <jcs@vectorized.io>