r/consensus: using last snapshot index to determine log start offset
Log eviction mechanism truncate log prefix if requested data retention
limit is breached. Log eviction have to be coordinated between storage
and raft since raft implementation needs a snapshot that will contain
metadata (offset, configuration)  related with the part of a log being
evicted. In redpanda we achieve that by notifying `raft::consensus`
about eviction that is required by `storage::log` cleanup policy. This
eviction notification triggers persisting a snapshot in raft and updates
configuration manager. In next pass (by default 10 seconds later) log is
safe to be evicted since raft has all metadata in the snapshot. In the
period of time between taking a snapshot and actually evicting log
offset translation layer may receive a request to translate offset for
which there are no longer information in `raft::configuration_manager`
since it was truncated during taking a snapshot. There is no information
allowing us to translate this offset correctly.

This commit changes the source of truth for the
`partition::start_offset`. When snapshot is taken in raft log it mean
that first offset available in log is exactly `last_snapshot_offset +
1`. This allow us to prevent `partition::start_offset` accessing offsets
for which translation delta can not be calculated. Change to
`raft::start_offset` is safe since it will always be update BEFORE
`log::start_offset`.

Signed-off-by: Michal Maslanka <michal@vectorized.io>