archival: Rewrite ntp_archiver
The coroutine version of the upload loop triggered a compiler bug
with clang 12 and ARM. This commit moves from using coroutines to
continuations to prevent this.

The logic is the same. To make it readable with continuations the
code was refactored a bit. The logic was split into several methods
and the state that was maintained on the stack in the previous version
is now passed around using 'scheduled_upload' structures.