test: reduce producer throughput
Even though we stop the e2e producers after 100K records are produced it
appears that producing is fast enough that by the time we notice and
stop a much more significnat amount of data is produced (e.g. >2M) and
when this happens the test client is slowing down because the consumer
logs are 100's of MBs in size and are also read up stored in memory
in inefficient python structures.

Signed-off-by: Noah Watkins <noah@vectorized.io>