cluster: handle partition deletions during update
This is a very rare condition -- ordinarily if a topic
is deleted during an update, the topic is torn down before
an update_finished can be generated.  However, it's not
obviously guaranteed to be that way, so also handle
deletion messages if they appear between and update
and an update_finished delta.

Signed-off-by: John Spray <jcs@vectorized.io>