tests: reinstate tests that were silently not running
Ducktape will ignore any test classes it can't import.  That
includes modules that don't import because of undefined names.

In a3a73c8eb3, `cluster` decorators were added to some functions
in modules where the name `cluster` was not in scope.  Ducktape

Ducktape logs these but does not fail the overall suite for it.

Signed-off-by: John Spray <jcs@vectorized.io>