out: add PrintStructFields (&test), PrintSection, PrintSectionHeader
---

Updating kmsg previously broke the output on a few commands. This is
because the output was using `tw.Print`, which accepts an arbitrary
amount of interfaces. The input was not type checked, kmsg (currently)
has no compatibility guarantee on types for updates, and the Kafka
protocol slightly changed (for non-OO languages).

One way to fix this is to add tests ensuring that every argument passed
to `tw.Print` is the argument we expect. This would be useful but
verbose, and there is an alternative.

The new PrintStructFields accepts a struct and prints all public fields
in the struct. The implementation is pretty simple, and importantly, any
call site passes a struct with typed fields to be printed. We now
control, at the call site, which types we are printing. This means we
have compile time checking on the types of arguments, so nothing in the
future can break the types.

This is a decent middle ground that avoids testing every argument all of
the time, while still having high guarantees on correctness.

---

Lastly, PrintSection and PrintSectionHeader are abstracted bits of code
that have been used in a few places already and will be used more.