go.{mod,sum}: add github.com/twmb/types
This is a small helper package that I plan to use in just about anything
that will print to the terminal going forward. In short, we can now
deeply sort anything so that we can have nice output all of the time.

The types.Sort function follows relatively obvious sorting, but most
usefully, sorts any type, and any field within any type.