acl: add common2{,_test}.go, to be renamed common.go
This is essentially the brains of the incoming acl conversion from
sarama to franz-go.

This new "acls" type extracts all of the logic that currently exists in
create/delete/list. Now that we no longer depend on sarama, we can
better construct **single** requests for creation/deletion.

This redux introduces the concept of specifying multiple resources at a
time. This is how kafka-acls.sh currently works.

Everything about this new type is heavily tested. The next few commits
will be converting the existing commands to using this new type.