acl list: convert to franz-go
Stealing the output from the prior commit, s/deletion/describing/:
```
This switches to using the new acls type. This conversion is backwards
compatible, but hides old flags and allows for the new multi-resource
behavior.

This also uses the new `tw.PrintStructFields` function to ensure our
output is typed checked.

The tests have been removed because the acl creation actually is heavily
tested in common2_test.go currently, and the response is compile-time
type checked, there is not much left to actually directly test.
```

In addition, this also clears up some now unused variables in the acl.go
file. The prior three commits **have** lost the ability to deduce
brokers from a running docker container; this ability has slowly been
lost in the conversion to ParamsFromConfig.