wasm: convert to franz-go, simplify
This commit is somewhat of a rewrite of the existing wasm code,
introducing a lot of simplifications and in the process halving the code
size. The create & remove tests have been removed because the logic is
now much simpler (and we do not use the kafka mocks). The generate test
has actually been strengthened, even though it looks opposite.

For deploying and removing, the code previously checked whether the
coproc topic existed before trying to create it. We can just always
directly issue a create topics call, this is simpler and avoids any
TOCTUO questions. We also simplify record creation and some other common
functions / code.

We remove specifying the number of replicas: redpanda previously
defaulted to 1, but it is changing to default to 3 if the cluster has
three members. We should just use the cluster default.

The generate code now checks if *any* file exists before writing any
file. This ensures that we do not create three files and fail on the
fourth. The test now also enforces this. Three of the unit tests have
been merged into one, because the three all tested the same thing (and
in this successful unit test, we added the check for files not
existing).

The template directory getter functions now are not named GetXyz.
Following the Go convention, getters should not be prefixed with Get.

The output for deploy / remove now indicates that the deploy/remove was
successful, rather than just printing that some message was produced at
a given offset.