cluster: Modify topic_configuration
Add 'recovery' and 'shadow_indexing' fields to 'topic_properties'. The
fields are used to represent topic configuration parameters
'x-redpanda-recovery', 'x-redpanda-archival', and
'x-redpanda-shadowindexing'.

Also, modify serialization of 'topic_configuration' since the properties
are serialized as part of the larger struct.

Introduce the version fields for the struct that goes first. The version
is a negative value which takes the same place as the size prefix of the
namespace field in the binary representation of the original struct.

The decoder checks this fields and if it's positive, then the struct is
decoded using the old procedure. If it's negative it's treated as a
version field and new decoding algorithm is used. To check the version
fields the decoder uses 'peek' method of the 'iobuf_parser'.

Add unt-tests that check that updated struct is binary compatible with
previous version.