id_allocator: remove state_machine based impl.
The problem with state_machine is that we only know its true state
after we  replicated a command and  waited until the state machine
caught up with it.

As a result it's hard remain consistent when we optimize the process
by updating the state in memory and only occasionally flushing it to
log.

Removing the state_machine based implementation to free room for the
persisted_stm which makes it easier to do such optimizations.

Leaving on disk data structures for backward compatibility:

  - sequence_id
  - allocation_cmd
  - prepare_truncation_cmd
  - execute_truncation_cmd

(cherry picked from commit c154e417f60ac1bdc4bb74971c1780a91b7f157c)