schema_registry: Sanitize type names and namespaces
The name field for complex type (record, union, fixed), should
have a leading dot removed iff it's the only one. Otherwise it
should be split on the last dot into a name and namespace.

The name of a type referred to in a union should have the leading
dot removed iff it's the only one.

Signed-off-by: Ben Pope <ben@vectorized.io>
(cherry picked from commit 4aac62cfae6f8ac29d30518c943378ff60455996)