cluster: fixed semantics of create partitions api
Previous implementation of create partitions API assumed that requested
number of partitions will be added to current topic partition count.
This assumption is incorrect since requested number of partitions is the
new total number of partitions for given topic. Adjusted implementation
and tests to account for that change in semantics.

Fixes: #2393

Signed-off-by: Michal Maslanka <michal@vectorized.io>