tests/prefix_truncate_recovery_test: correct num_nodes definition
It would appear that the `@cluster` decorator will silently
ignore unknown arguments, and the result is that the test runs
but without a defined node count, and ducktape logs a warning
that the test is using the entire cluster.

This was probably increasing overall runtime by cutting out
parallelism while this test ran.

Signed-off-by: John Spray <jcs@vectorized.io>