tests: remove produce_bench_test
This looks like a redpanda equivalent of the test by the same
name in the kafka tree.  However it doesn't run in CI and
can't unless we change things to make `kafkatest` module
available.

Borrowing some Trogdor test might be a good idea at some
point, but keeping this unused draft in our tree is not
useful.  Removing this lets us put the check in place
to catch tests with import errors.

Signed-off-by: John Spray <jcs@vectorized.io>