compaction: decompress batch in kafka layer
There are two reasons why we want to decompress incoming message batch:

 - to update compaction index
 - to validate the request

So far we were ignoring the validation part and used to decompress the batch
only in the storage layer. But it's too deep and too late for validation!

Moving the decompression up the stack in a place suitable for validation.