rpk topic produce: redo for franz-go
This rewrites the produce code to be specific to franz-go. The new code
does effectively no work other than parsing options and beginning a
produce loop; parseKVs and the kgo.RecordReader are very heavily tested.
Since we have integration tests for producing, and because franz-go has
no mock client, the tests ensuring a produce actually happens are
removed.

The produce command can now write many records at a time, rather than
one.

Changes:

- jvm-partitioner is deprecated; it is now the default. There's no
reason for a user of a CLI to need to worry about this detail.

- num has been deprecated; if a user wants to repeat messages, they can
call the produce command many times. This flag is less important now
that the produce command can accept and write many records.

- timestamp has been deprecated; it was not really valid anyway.
Clients internally override timestamps when producing.

- The default parse format has been changed to _not_ write a trailing
newline.

The three flag deprecations are soft deprecations; we hide the flag and
using it does nothing, but the command will still succeed.